fastapi_movietheatre/fastapi-solution/requirements.txt
fastapi==0.111.0
elasticsearch[async]==8.13.2
redis==5.0.4
gunicorn==23.0.0

httpx==0.28.1
python-jose==3.3.0

opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-jaeger==1.21.0